Split route for Prediction Targets Admin
