+++ src/app/(payload)/admin/importMap.js
@@ -2,10 +2,12 @@ import { default as default_aa936f6d4bfcfa210342637718f4c629 } from '@/component
 import { default as default_ad44889f58df4e99210146c51158a6ba } from '@/components/Pages/RaceList'
 import { default as default_d5a1575951a32156bf5f24180b347a19 } from '@/components/Pages/RaceSettings'
 import { default as default_82cb54cea17942e21ee8956c92300c78 } from '@/components/Pages/PredictionTargets'
+import { default as default_a0a77058b8a476d297b9f7a3fedbf647 } from '@/components/Pages/PredictionTargetsAdmin'
 
 export const importMap = {
   "@/components/AfterNavLinks#default": default_aa936f6d4bfcfa210342637718f4c629,
   "@/components/Pages/RaceList#default": default_ad44889f58df4e99210146c51158a6ba,
   "@/components/Pages/RaceSettings#default": default_d5a1575951a32156bf5f24180b347a19,
-  "@/components/Pages/PredictionTargets#default": default_82cb54cea17942e21ee8956c92300c78
+  "@/components/Pages/PredictionTargets#default": default_82cb54cea17942e21ee8956c92300c78,
+  "@/components/Pages/PredictionTargetsAdmin#default": default_a0a77058b8a476d297b9f7a3fedbf647
 }
