+++ src/components/Pages/PredictionTargetsAdmin/container/index.tsx
@@ -1,16 +1,15 @@
 'use client'
 
+import StatusField from '@/components/Common/StatusField'
 import { Pagination, Table } from '@payloadcms/ui'
 import React from 'react'
-import StatusField from '../component/StatusField'
 
-export const PredictionTargetsAdminContent = () => {
+export const PredictionTargetsAdminContainer = () => {
   const renderCells = <T,>(data: T[], getContent: (item: T, idx: number) => React.ReactNode) =>
     data.map((item, idx) => <div key={idx}>{getContent(item, idx)}</div>)
 
   return (
     <div className="align-middle">
-      <h1 className="mb-5 font-bold">レース一覧</h1>
       <Table
         data={races}
         columns={[
