+++ src/libs/enums.ts
@@ -3,6 +3,7 @@ export enum ROUTE_CONFIG {
   RACE_LIST = '/race-list',
   RACE_SETTINGS = '/race-list/race-settings',
   PREDICTION_TARGET = '/prediction-target',
+  PREDICTION_TARGET_ADMIN = '/prediction-target-admin',
 }
 
 // Apply when navigate
@@ -20,4 +21,5 @@ export enum ROUTE_NAVIGATE {
   RACE_LIST = '/admin/race-list',
   RACE_SETTINGS = '/admin/race-list/race-settings',
   PREDICTION_TARGET = '/admin/prediction-target',
+  PREDICTION_TARGET_ADMIN = '/admin/prediction-target-admin',
 }
